==> 15_markdown_V2.Rmd <==
---
title: "markdown_test"
output: html_document
date: "2024-01-08"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, include = FALSE, warning=FALSE)
```

```{r}
pacman::p_load(tidyverse, haven, jtools, lme4, lmerTest, sjPlot)
rm(list = ls())
```

## Cross-sectional opposition to immigration modelling

Dependent variable and model:  

* Question: 'Some people think that the UK should allow *many more* immigrants to come to the UK to live and others think that the UK should allow *many fewer* immigrants. Where would you place yourself on this scale?'. Responses on an 11-point scale from 0 (should allow many more immigrants) to 10 (should allow far fewer immigrants) (note: original scale had 0 = far fewer immigrants, but scale has been reversed for analysis)  
* Nested multilevel linear model with random intercepts for local authorities, with cross-level interactions between housing affordability and a) social renting and b) homeownership  
* Affordability measured via the ratio of median house prices to median earnings in a local authority; as such, a higher number indicates less affordable housing  
* Large amounts of missing data for two key individual level predictors: university educated, and income. As such, findings are presented for three models. The 'Main' model omits incomes and includes the age at which a respondent left education as a proxy for university education, thus maximising possible sample size. The 'Uni' model uses university education as a predictor instead of age left education, with missing observations removed via listwise deletion. And the 'Income' model includes income as a predictor, with missing values imputed via a Mincer regression model  

Key findings:  

* The association between affordability and opposition to immigration is conditional on tenure
* A positive interaction exists between social renting and unaffordability i.e. social renters in less affordable areas express stronger opposition to immigration  
* A negative interaction exists between homeownership and unaffordability i.e. homeowners in less affordable areas express stronger support for immigration  
* Findings are consistent across the Main, Uni and Income models  
* Robustness checks include: using log transformation of affordability as independent variable; using log of house prices as independent variable; inclusion of dummy variable for region to ensure it's not just a 'London effect';   

```{r, include=TRUE}
# plot
readRDS("working/markdown_viz/coef_plot_immi.RDS")
```

The figure below displays the interaction terms. It also shows that once controlling for the confounding influence of the percentage of people with degrees in a local authority, for many tenures there is a positive association between high house prices and right wing political attitudes.  

```{r, include = TRUE}
readRDS("working/markdown_viz/int_plot.RDS")
```


## Cross-sectional immigration control modelling

Dependent variable and model:  

* Question: "As Britain negotiates its future relationship with the EU, is it more important for the UK government to protect Britain's access to the single market or to gain full control of immigration?" Responses given on an 11-point scale from 0 (access to the single market) to 10 (control immigration)  
* Nested multilevel linear model with random intercepts for local authorities, with cross-level interactions between housing affordability and a) social renting and b) homeownership  

Key findings:  

* Positive interaction between social renting and affordability i.e. social renters in less affordable areas more likely to want to control immigration as opposed to access the single market  
* No interaction with homeownership  
* Interaction between social renting and affordability robust to the same robustness checks as described above 

```{r, include=TRUE}
# plot
readRDS("working/markdown_viz/coef_plot_eu.RDS")
```

```{r, include = TRUE}
readRDS("working/markdown_viz/int_plot_eu.RDS")
```


## Cross-sectional racial equality modelling

Dependent variable and model:  

* Opposition to racial equality measured as a binary indicator (1 = racial equality has "gone too far" or "gone much too far")  
* Nested multilevel logit model with random intercepts for local authorities, with cross-level interactions between housing affordability and a) social renting and b) homeownership  

Key findings:  

* No effect of housing affordability, or the interaction terms
* Homeowners more likely to think racial equality has gone too far than other tenures  

```{r, include=TRUE}
readRDS("working/markdown_viz/coef_plot_equality.RDS")
```

## Table of cross-sectional modelling results

```{r, include=TRUE, results = 'asis'}
immi_int <- readRDS("working/markdown_data/immi_int.RDS")
eu_int <- readRDS("working/markdown_data/immig_eu_int.RDS")
equal_int <- readRDS("working/markdown_data/equal_int.RDS")

tab_model(immi_int, eu_int, equal_int, p.style = "stars", show.aic = T, 
          terms = c("social_housing", "homeowner", 
                    "private_renting", "affordability", 
                    "social_housing.affordability", 
                    "homeowner.affordability"),
          pred.labels = c("Social housing", "Homeowner", "Private renter",
                          "Affordability",
                          "Social housing * Affordability", 
                          "Homeowner * Affordability"),
          dv.labels = c("Fewer immigrants", "Control immigration", "Racial equality too far"))
```


## Longitudinal anti-immigration modelling

Data and modelling strategy:

* Six waves of the British Election Study (BES) ranging from 2016-2021, filtered to respondents who participated in at least three waves  
* Random effect within-between (REWB) model with two-way cross-classified structure i.e. random intercept for respondent and random intercept for local authority  
* Cross-level interaction between tenure and the average affordability of a local authority across years (referred to as 'affordability between')  
* The 'within' effect is the difference between affordability in an authority in a given year and average affordability in an authority across all years (i.e. 'affordability in year i authority j' - 'mean affordability in authority j'). This is 'affordability within', and tests for a *geotropic effect* i.e. the effect of living in an area of declining affordability irrespective of tenure  

Key findings:  

* Association of between affordability with opposition to immigration is again contingent upon tenure, with the direction of the interactions consistent with the cross-sectional modelling  
* Variation in affordability within local authorities not associated with the outcome variable; therefore, no evidence of geotropic effect  
* In sum, evidence suggests that differences *between* authorities in terms of affordability are more important for political attitudes than changes occurring *within* authorities over time  
* Findings robust to robustness checks described above  

```{r, include=TRUE, results = 'asis'}
immi_long <- readRDS("working/markdown_data/immi_int_long.RDS")

tab_model(immi_long, p.style = "stars", show.aic = T, 
          terms = c("social_housing", "homeowner", 
                    "private_renting", "affordability_mean",
                    "social_housing.affordability_mean",
                    "homeowner.affordability_mean",
                    "affordability_within"),
          pred.labels = c("Social housing", "Homeowner", "Private renter",
                          "Affordability between",
                          "Social housing * Affordability between",
                          "Homeowner * Affordability between",
                          "Affordability within"),
          dv.labels = c("Fewer immigrants (longitudinal)"))
```
